Trusted services (keystore, cache, others?) are now keyed off the request apps 'project' property rather than 'guid'.  The value of 'project' defaults to 'guid' (which itself defaults to 'host').
Apps and appfetchers can change this value as long as the new value is just further up the directory hierarchy from the host value.  For example, the app homepage.www.trunk.svn.freebase-site.googlecode.dev can set it's project value to freebase-site.googlecode.dev... but not to acre.googlecode.dev.

Currently the googlecode appfetcher will set project to be the google code project (examples above)... but none of the other appfetchers have been modified.  So the only backwards-compatibility issues are with googlecode apps.